refactor: move offchain code
