+++ nightfall-client/src/services/burn.mjs
@@ -9,10 +9,9 @@ import {
 import gen from 'general-number';
 import Transaction from '@polygon-nightfall/common-files/classes/transaction.mjs';
 import { compressProof } from '@polygon-nightfall/common-files/utils/curve-maths/curves.mjs';
-import { clearPending } from './commitment-storage.mjs';
+import { clearPending, saveExtendedTransaction } from './commitment-storage.mjs';
 import { getCommitmentInfo } from '../utils/getCommitmentInfo.mjs';
 import { computeCircuitInputs } from '../utils/computeCircuitInputs.mjs';
-import { submitTransaction } from '../utils/submitTransaction.mjs';
 import { ZkpKeys } from './keys.mjs';
 
 const { VK_IDS } = config;
@@ -124,7 +123,7 @@ async function burn(burnParams) {
     const rawTransaction = await shieldContractInstance.methods
       .submitTransaction(Transaction.buildSolidityStruct(transaction))
       .encodeABI();
-    await submitTransaction(
+    await saveExtendedTransaction(
       transaction,
       commitmentsInfo,
       compressedZkpPublicKey,
