+++ nightfall-client/src/services/transform.mjs
@@ -12,10 +12,9 @@ import {
 import { compressProof } from '@polygon-nightfall/common-files/utils/curve-maths/curves.mjs';
 import gen from 'general-number';
 import { Commitment, Transaction } from '../classes/index.mjs';
-import { clearPending } from './commitment-storage.mjs';
+import { clearPending, saveExtendedTransaction } from './commitment-storage.mjs';
 import { getCommitmentInfo } from '../utils/getCommitmentInfo.mjs';
 import { computeCircuitInputs } from '../utils/computeCircuitInputs.mjs';
-import { submitTransaction } from '../utils/submitTransaction.mjs';
 import { ZkpKeys } from './keys.mjs';
 
 const { VK_IDS } = config;
@@ -176,7 +175,7 @@ async function transform(transformParams) {
       .submitTransaction(Transaction.buildSolidityStruct(transaction))
       .encodeABI();
 
-    await submitTransaction(
+    await saveExtendedTransaction(
       transaction,
       commitmentInfo,
       compressedZkpPublicKey,
