+++ nightfall-client/src/services/withdraw.mjs
@@ -17,9 +17,8 @@ import {
 } from '@polygon-nightfall/common-files/utils/worker-calls.mjs';
 import { Transaction } from '../classes/index.mjs';
 import { computeCircuitInputs } from '../utils/computeCircuitInputs.mjs';
-import { clearPending } from './commitment-storage.mjs';
+import { clearPending, saveExtendedTransaction } from './commitment-storage.mjs';
 import { getCommitmentInfo } from '../utils/getCommitmentInfo.mjs';
-import { submitTransaction } from '../utils/submitTransaction.mjs';
 import { ZkpKeys } from './keys.mjs';
 
 const { VK_IDS } = config;
@@ -137,7 +136,7 @@ async function withdraw(withdrawParams) {
     const rawTransaction = await shieldContractInstance.methods
       .submitTransaction(Transaction.buildSolidityStruct(transaction))
       .encodeABI();
-    await submitTransaction(
+    await saveExtendedTransaction(
       transaction,
       commitmentsInfo,
       compressedZkpPublicKey,
